sw_services: xilplmi: Fixes for MISRA C violation Rule 12.2
MISRA C Rule 12.2 States: The right hand operand of a shift operator shall
lie in the range zero to one less than the width in bits of the essential
type of the left hand operand

This patch fixes MISRA C Rule 12.2 violation by casting constants with higher
type when shifting more than 7 bits


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilplmi/src/xplmi_dma.h
@@ -46,13 +46,13 @@ extern "C" {
 /** DMA XFER flags */
 #define XPLMI_SRC_CH_AXI_FIXED		(0x1U)
 #define XPLMI_DMA_SRC_NONBLK		(0x1U << 1U)
-#define XPLMI_DST_CH_AXI_FIXED		(0x1U << 16U)
-#define XPLMI_DMA_DST_NONBLK		(0x1U << 17U)
+#define XPLMI_DST_CH_AXI_FIXED		((u32)0x1U << 16U)
+#define XPLMI_DMA_DST_NONBLK		((u32)0x1U << 17U)
 #define XPLMI_PMCDMA_0				(0x100U)
 #define XPLMI_PMCDMA_1				(0x200U)
 #define XPLMI_DMA_SRC_NPI			(0x4U)
 #define XPLMI_DMA_DST_TYPE_SHIFT	(18U)
-#define XPLMI_DMA_DST_TYPE_MASK		(0x3U << 18U)
+#define XPLMI_DMA_DST_TYPE_MASK		((u32)0x3U << 18U)
 #define XPLMI_READ_AXI_FIXED		(0x1U)
 
 /* SSS configurations and masks */
